feat: 포스트 조회 pagination 구현
- 포스트 정렬기준 enum 추가

+++ src/main/java/com/nexters/phochak/dto/request/PostPageRequestDto.java
@@ -8,12 +8,12 @@ import javax.validation.constraints.NotNull;
 public class PostPageRequestDto {
     @NotNull
     private Long lastId;
-    private Integer pageSize;
+    private Long pageSize;
     @NotNull
     private String postSortCriteria;
     private Long lastCriteriaValue;
 
-    public PostPageRequestDto(Long lastId, Integer pageSize, String postSortCriteria, Long lastCriteriaValue) {
+    public PostPageRequestDto(Long lastId, Long pageSize, String postSortCriteria, Long lastCriteriaValue) {
         this.lastId = lastId;
         this.pageSize = pageSize;
         this.postSortCriteria = postSortCriteria;
